Fix magit loading
Since switching to Elpaca (7969e0a), a couple of things broke; one of
which was undesirable loading behaviour of magit.  This meant that the
MIT snippet, which relies on `magit-git-string`, does not work unless
a magit buffer is opened, and when opening a magit buffer, it was
slow.  In reality, we want to load magit as soon as possible.