Stop magit from loading before it's ready
In a (silly) attempt at eagerly loading magit so that we have access
to all of its useful features, we `require` magit right after
instantiating it with `use-package` (73f1ae2).  However, `use-package`
with `:defer t` is lazy, and `require` (thereby adding it to
`load-path` is eager, which ultimately caused issues.  @progfolio
aptly pointed this out.

Vaguely addresses #7